Moving MediaStreamSignaling logic into PeerConnection.
This needs to happen because in the future, m-lines will be offered
based on the set of RtpSenders/RtpReceivers, rather than the set of
tracks that MediaStreamSignaling knows about.

Besides that, MediaStreamSignaling was a "glue class" without
a clearly defined role, so it going away is good for other
reasons as well.

Review URL: https://codereview.webrtc.org/1393563002

Cr-Original-Commit-Position: refs/heads/master@{#10268}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 97c392935411398b506861601c82e31d95c591f0